feat(frontend): Add some layers to a map
I've created a deck.gl based map and added a few layers to play with.
Some are commented out as I only want to focus on one at a time, but
others are free to show all at once, or anything they wish really.

To get this working, the user will need to register with mapbox and
get their own personal `MAPBOX_TOKEN`. Just remember not to commit
that, as I have done in the past...